// fe/src/lib/api/generated/model/referralActionResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { ReferralActionTypeValue } from './referralActionTypeValue';

export interface ReferralActionResponse {
  action_type: ReferralActionTypeValue;
  created_at: string;
  id: string;
  occurred_at: string;
  remarks: string;
}